+++ integrations/gitlab-v2/gitlab/webhook/webhook_processors/_gitlab_abstract_webhook_processor.py
@@ -19,11 +19,14 @@ class _GitlabAbstractWebhookProcessor(AbstractWebhookProcessor):
         return True
 
     async def should_process_event(self, event: WebhookEvent) -> bool:
-        event_name = str(
-            event.payload.get("event_name") or event.payload.get("event_type")
+        event_identifier = (
+            event.payload.get("event_name")
+            or event.payload.get("event_type")
+            or event.payload.get("object_kind")
         )
         return bool(
-            event.headers["x-gitlab-event"] in self.hooks and event_name in self.events
+            event.headers["x-gitlab-event"] in self.hooks
+            and event_identifier in self.events
         )
 
     async def validate_payload(self, payload: EventPayload) -> bool: